+++ Concurrency.Implementation/Logging/Simple2PCLoggingProtocol.cs
@@ -13,7 +13,7 @@ namespace Concurrency.Implementation.Logging
         IKeyValueStorageWrapper logStorage;
         String grainType;
         Guid grainPrimaryKey;
-        long sequenceNumber;
+        int sequenceNumber;
 
         public Simple2PCLoggingProtocol(String grainType, Guid grainPrimaryKey) {
             this.grainType = grainType;
@@ -24,9 +24,9 @@ namespace Concurrency.Implementation.Logging
             //logStorage = new DynamoDBStorageWrapper(grainType, grainPrimaryKey);
         }
 
-        private long getSequenceNumber()
+        private int getSequenceNumber()
         {
-            long returnVal;
+            int returnVal;
             lock (this)
             {
                 returnVal = sequenceNumber;
@@ -34,44 +34,44 @@ namespace Concurrency.Implementation.Logging
             }
             return returnVal;
         }
-        async Task ILoggingProtocol<TState>.HandleBeforePrepareIn2PC(long tid, Guid coordinatorKey, HashSet<Guid> grains)
+        async Task ILoggingProtocol<TState>.HandleBeforePrepareIn2PC(int tid, Guid coordinatorKey, HashSet<Guid> grains)
         {
             var logRecord = new LogParticipant(getSequenceNumber(), grainPrimaryKey, tid, grains);
             await logStorage.Write(BitConverter.GetBytes(logRecord.sequenceNumber), Helper.serializeToByteArray<LogParticipant>(logRecord));
         }
 
-        async Task ILoggingProtocol<TState>.HandleOnAbortIn2PC(ITransactionalState<TState> state, long tid, Guid coordinatorKey)
+        async Task ILoggingProtocol<TState>.HandleOnAbortIn2PC(ITransactionalState<TState> state, int tid, Guid coordinatorKey)
         {            
             var logRecord = new LogFormat<TState>(getSequenceNumber(), LogType.ABORT, coordinatorKey, tid);            
             await logStorage.Write(BitConverter.GetBytes(logRecord.sequenceNumber), Helper.serializeToByteArray<LogFormat<TState>>(logRecord));
         }
 
-        async Task ILoggingProtocol<TState>.HandleOnCommitIn2PC(ITransactionalState<TState> state, long tid, Guid coordinatorKey)
+        async Task ILoggingProtocol<TState>.HandleOnCommitIn2PC(ITransactionalState<TState> state, int tid, Guid coordinatorKey)
         {
             var logRecord = new LogFormat<TState>(getSequenceNumber(), LogType.COMMIT, coordinatorKey, tid);
             await logStorage.Write(BitConverter.GetBytes(logRecord.sequenceNumber), Helper.serializeToByteArray<LogFormat<TState>>(logRecord));
             
         }
 
-        async Task ILoggingProtocol<TState>.HandleOnPrepareIn2PC(ITransactionalState<TState> state, long tid, Guid coordinatorKey)
+        async Task ILoggingProtocol<TState>.HandleOnPrepareIn2PC(ITransactionalState<TState> state, int tid, Guid coordinatorKey)
         {
             var logRecord = new LogFormat<TState>(getSequenceNumber(), LogType.PREPARE, coordinatorKey, tid, state.GetPreparedState(tid));
             await logStorage.Write(BitConverter.GetBytes(logRecord.sequenceNumber), Helper.serializeToByteArray<LogFormat<TState>>(logRecord));            
         }
 
-        async Task ILoggingProtocol<TState>.HandleOnCompleteInDeterministicProtocol(ITransactionalState<TState> state, long bid, Guid coordinatorKey)
+        async Task ILoggingProtocol<TState>.HandleOnCompleteInDeterministicProtocol(ITransactionalState<TState> state, int bid, Guid coordinatorKey)
         {
             var logRecord = new LogFormat<TState>(getSequenceNumber(), LogType.DET_COMPLETE, coordinatorKey, bid, state.GetCommittedState(bid));
             await logStorage.Write(BitConverter.GetBytes(logRecord.sequenceNumber), Helper.serializeToByteArray<LogFormat<TState>>(logRecord));
         }
 
-        async Task ILoggingProtocol<TState>.HandleOnPrepareInDeterministicProtocol(long bid, HashSet<Guid> grains)
+        async Task ILoggingProtocol<TState>.HandleOnPrepareInDeterministicProtocol(int bid, HashSet<Guid> grains)
         {
             var logRecord = new LogParticipant(getSequenceNumber(), grainPrimaryKey, bid, grains);
             await logStorage.Write(BitConverter.GetBytes(logRecord.sequenceNumber), Helper.serializeToByteArray<LogParticipant>(logRecord));
         }
 
-        async Task ILoggingProtocol<TState>.HandleOnCommitInDeterministicProtocol(long bid)
+        async Task ILoggingProtocol<TState>.HandleOnCommitInDeterministicProtocol(int bid)
         {
             var logRecord = new LogFormat<TState>(getSequenceNumber(), LogType.DET_COMMIT, grainPrimaryKey, bid);
             await logStorage.Write(BitConverter.GetBytes(logRecord.sequenceNumber), Helper.serializeToByteArray<LogFormat<TState>>(logRecord));
